Move StatusRenderer
