+++ src/frontend/src/pages/Index/Playground.tsx
@@ -8,7 +8,7 @@ import { ApiFormProps } from '../../components/forms/ApiForm';
 import { PlaceholderPill } from '../../components/items/Placeholder';
 import { StylishText } from '../../components/items/StylishText';
 import { ModelType } from '../../components/render/ModelType';
-import { StatusRenderer } from '../../components/renderers/StatusRenderer';
+import { StatusRenderer } from '../../components/render/StatusRenderer';
 import {
   createPart,
   editPart,
